TIP-706: Move field sorter to dedicated directory
