+++ src/Pim/Bundle/CatalogBundle/Elasticsearch/Sorter/Attribute/MetricSorter.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Elasticsearch\Sorter\Attributes;
+namespace Pim\Bundle\CatalogBundle\Elasticsearch\Sorter\Attribute;
 
 /**
  * Metric sorter for an Elastic search query
