+++ src/Pim/Bundle/CatalogBundle/spec/Elasticsearch/Sorter/Attribute/TextAreaSorterSpec.php
@@ -3,13 +3,12 @@
 namespace spec\Pim\Bundle\CatalogBundle\Elasticsearch\Sorter\Attributes;
 
 use Pim\Bundle\CatalogBundle\Elasticsearch\SearchQueryBuilder;
-use Pim\Bundle\CatalogBundle\Elasticsearch\Sorter\Attributes\TextAreaSorter;
+use Pim\Bundle\CatalogBundle\Elasticsearch\Sorter\Attribute\TextAreaSorter;
 use PhpSpec\ObjectBehavior;
 use Pim\Component\Catalog\Exception\InvalidDirectionException;
 use Pim\Component\Catalog\Model\AttributeInterface;
 use Pim\Component\Catalog\Query\Sorter\AttributeSorterInterface;
 use Pim\Component\Catalog\Query\Sorter\Directions;
-use Prophecy\Argument;
 
 class TextAreaSorterSpec extends ObjectBehavior
 {
